# openapi.yaml
openapi: 3.1.0
info:
  title: IoT and IM Microservices API
  description: |
    Formal API specification matching the provided Python microservices:
    - **IMTelegram Service**: Port 5000 (Prefix: /IMTelegram)
    - **IMLine Service**: Port 5001 (Prefix: /IMLine)
    - **ESP32 Service**: Port 5002/5010 (Prefix: /ESP32)
    - **Raspberry Pi Service**: Port 5003/5011 (Prefix: /Pi)
  version: 2.0.2
  contact:
    name: Rocer52
    email: batte9306@gmail.com
servers:
  - url: http://localhost:5000
    description: IM Telegram Service
  - url: http://localhost:5001
    description: IM Line Service
  - url: http://localhost:5002
    description: ESP32 Service (IoT)
  - url: http://localhost:5003
    description: Raspberry Pi Service (IoT)
...
paths:
...
  /IMTelegram/webhook:
    post:
      tags:
        - Telegram
      summary: Handle Telegram Webhook
      description: Endpoint to receive webhook events from Telegram.
      operationId: handleTelegramWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramWebhookRequest'
      responses:
        '200':
          description: Processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
  /IMTelegram/SendMsg:
    get:
      tags:
        - Telegram
      summary: Send Telegram Message
      description: Manually send a message to a specific Telegram chat/user.
      operationId: sendTelegramMessage
      parameters:
        - $ref: '#/components/parameters/chat_id'
        - $ref: '#/components/parameters/message'
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
  /IMTelegram/SendGroupMessage:
    get:
      tags:
        - Telegram
      summary: Broadcast to Device Group (Telegram)
      description: Send a message to Telegram users bound to a specific device.
      operationId: sendTelegramGroupMessage
      parameters:
        - $ref: '#/components/parameters/device_id'
        - $ref: '#/components/parameters/message'
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Messages sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'
  /IMTelegram/SendAllMessage:
    get:
      tags:
        - Telegram
      summary: Broadcast to All Users (Telegram)
      description: Send a message to all users known to the Telegram service.
      operationId: sendTelegramAllMessage
      parameters:
        - $ref: '#/components/parameters/message'
        - $ref: '#/components/parameters/user_id'
        - $ref: '#/components/parameters/bot_token'
      responses:
        '200':
          description: Messages sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicResponse'

  # ==============================
  # Line Microservice Paths
  # ==============================
...
components:
  schemas:
    BasicResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: "Message sent"
      required:
        - ok
        - message
...
    TelegramWebhookRequest:
      type: object
      properties:
        message:
          type: object
          properties:
            text:
              type: string
              example: "turn on esp32_light_001"
            chat:
              type: object
              properties:
                id:
                  type: string
                  example: "123456789"
                type:
                  type: string
                  example: "private"
            from:
              type: object
              properties:
                id:
                  type: string
                  example: "123456789"
                username:
                  type: string
                  example: "user123"
          required:
            - text
            - chat
            - from
      required:
        - message
...
  parameters:
    device_id:
      name: device_id
      in: query
      description: Device ID for general queries
      required: true
      schema:
        type: string
        example: "esp32_light_001"
...
    chat_id:
      name: chat_id
      in: query
      description: Chat ID
      required: false
      schema:
        type: string
        example: "123456789"
    user_id:
      name: user_id
      in: query
      description: User ID
      required: false
      schema:
        type: string
        example: "123456789"
    message:
      name: message
      in: query
      description: Message content
      required: true
      schema:
        type: string
        example: "Hello!"
    bot_token:
      name: bot_token
      in: query
      description: Bot token
      required: false
      schema:
        type: string
        example: "123456789:ABCDEF_ExampleBotToken"
...